Fixed Weapon Switching
When switching weapons, there was an issue where it switches to another player prefab. This will make it harder for Elijah to implement the health system.

I updated the switching, where the player is just one object holding multiple weapons. I also adjusted the speed of the player's fire rate of the UZI.