Pause Branch Merged (Manually)
Manually merged the pause branch into this branch by taking bits of code and prefabs from the pause branch and implementing them here. This was due to merge conflicts and other issues occurring when making a pull request with the pause branch, as the pause branch was based off of a very old version of our code.

I have implemented the pause only on the HouseFirstFloor scene. Still have to add it to other scenes.